Added functions used by all controllers to base controller
